aioffice-python/src/requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.1
selenium==4.18.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.149.0